Fixes #23 : REST API getSize fails for very large investigations
The problem was in resolving the DataSelection for large investigations.
The code has been changed such that it can cope provided that the
maximum number of entities to return in ICAT is set to exceed both the
maximum number of datasets within an investigation and the maximum
number of datafiles within a dataset.